requirements.txt
flask==3.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
python-decouple==3.8
ldap3==2.9.1
cloudscraper